+++ .jrubyrc
@@ -1,0 +1,1 @@
+debug.fullTrace=true
