+++ .travis.yml
@@ -1,13 +1,15 @@
 language: ruby
 sudo: false
 cache: bundler
+before_script:
+  - gem update --system
 script: ./script/ci
 rvm:
-  - 2.3.6
-  - 2.4.2
-  - 2.5.0
+  - 2.4.4
+  - 2.3.7
+  - 2.5.1
+  - jruby-9.1.9.0
   - ruby-head
-  - jruby-9.1.13.0
   - jruby-head
 
 matrix:
