+++ Gemfile
@@ -1,7 +1,7 @@
 source 'https://rubygems.org'
 gemspec
 
-unless ENV['TRAVIS']
+unless ENV['CI']
   gem 'byebug', require: false, platforms: :mri
   gem 'yard',   require: false
 end
@@ -10,4 +10,3 @@ gem 'hanami-utils', '~> 1.2', require: false, git: 'https://github.com/hanami/ut
 gem 'haml'
 
 gem 'hanami-devtools', require: false, git: 'https://github.com/hanami/devtools.git'
-gem 'coveralls', require: false
